[IMP] tools: move json_default from date_utils to json
Historically, the `json_default` method was found in the `date_utils.py` file,
because it was used to consider `date` and `datetime` objects during serialization.

The place of the `json_default` method is in the `json.py` file.
Now, it can be used to serialise other types of object.
It therefore makes sense to move it.
