+++ odoo/tools/__init__.py
@@ -18,6 +18,7 @@ from .float_utils import *
 from .func import *
 from .i18n import format_list
 from .image import image_process
+from .json import json_default
 from .mail import *
 from .misc import *
 from .query import Query
